♻️ Refactor(#6): 도메인별 기능 분리, 코드 포맷팅
